finbuddy_app/requirements.txt
streamlit
pandas==2.0.3
pdfplumber
python-dateutil
regex
matplotlib
seaborn
squarify
openai
